requirements.txt
pandas
scikit-learn
tqdm
python-dotenv
pinecone-client
fastf1
